requirements.txt
openai
wolframalpha==5.0.0
langchain-community
python-dotenv
exa-py